requirements.txt
flask==2.2.5
Werkzeug==2.3.7
jinja2==3.1.2
flask-login==0.6.2
flask_migrate==4.0.4
WTForms==3.0.1
flask_wtf==1.2.1
flask-sqlalchemy==3.0.5
sqlalchemy==2.0.21
email_validator==2.0.0
flask-restx==1.1.0

python-dotenv==0.19.2

gunicorn==20.1.0
Flask-Minify==0.37

flask-dance==7.0.0
blinker==1.6.2

# flask_mysqldb
# psycopg2-binary

llama_index
google-cloud-storage
google-cloud-bigquery
openpyxl